// src/models/product.ts
export default interface Product {
    fabricante: number;
    produto: string;
    oferta_codigo: number;
    ordem_listagem: number;
    quantidade: number;
    dep_nome: string;
    data_acab: number;
    sec_nome: string;
    dep: string;
    vlr_normal: string;
    sec_amigavel: string;
    tmp_avaliacao_numero: number;
    tmp_ranking: number;
    sec: string;
    data_ini: number;
    tmp_avaliacao_nota: number;
    destaque_listagem: boolean;
    codigo: number;
    imagem: string;
    dep_amigavel: string;
    desconto: number;
    oferta_quantidade_vendida: number;
    data_fim: number;
    vlr_oferta: number;
}